+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Changed
+- Exception logging on lambda-style nested step was removed to avoid double-logging, since it should be logged on test level.
 
 ## [5.1.0-RC-12]
 ### Changed